+++ lib/src/modules/elss/elss_trc/screens/add_part_screen_trc.dart
@@ -1,5 +1,6 @@
 import 'package:core_widgets/core_widgets.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_trc/src/header/trc_header.dart';
 import 'package:provider/provider.dart';
 import '../l10n.dart';
 import '../providers/add_part_list_provider_trc.dart';
@@ -22,7 +23,7 @@ class AddPartScreenTrc extends StatelessWidget {
         var provider = AddPartListProviderTrc.of(innerContext);
 
         return Scaffold(
-          appBar: CshHeader(l10n.addPart),
+          appBar: TrcHeader(l10n.addPart),
           body: (provider.isPartListLoading)
               ? const Center(
                   child: SizedBox(