+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # TRC Releases
 
+# v1.2.0+2 (03-04-2023)
+
+### Improvements:
+
+* `QC-ELSS - change initial grade to current grade` [(SCM-2793)](https://reglobe.atlassian.net/browse/SCM-2793)
+* `QC-ELSS - Start showing quantity in suggested and added parts` [(SCM-2794)](https://reglobe.atlassian.net/browse/SCM-2794)
+* `QC-ELSS - Add retesting and rejection reasons` [(SCM-2874)](https://reglobe.atlassian.net/browse/SCM-2874)
+* `QC-ELSS - Mark parts as Not Required by default` [(SCM-2954)](https://reglobe.atlassian.net/browse/SCM-2954)
+
+
 # v1.0.0+1 (7-03-2023)
 
 ### Feature: